Carve FriendSlopSceneBuilder into per-system partial files
Splits the 1297-line monolith into a partial class spread across
five focused files under Assets/Scripts/Editor/Builders/. Zero
call-site changes - all methods stay private static on the same
FriendSlopSceneBuilder class, just physically located by responsibility.

New files:

- FriendSlopSceneBuilder.Materials.cs (99 lines): CreateMaterials,
  CreateMaterial, SetMaterial. The material library used by every
  other builder.
- FriendSlopSceneBuilder.PlayerPrefab.cs (128 lines): BuildPlayerPrefab.
  Adding a new player prefab variant or tweaking defaults touches one
  file instead of opening the giant.
- FriendSlopSceneBuilder.RoundManagerPrefab.cs (34 lines):
  BuildRoundManagerPrefab.
- FriendSlopSceneBuilder.MonsterPrefab.cs (74 lines): BuildMonsterPrefab.
- FriendSlopSceneBuilder.LootPrefab.cs (170 lines): BuildLootPrefabs,
  WarnAboutOrphanedLootPrefabs, LoadLootPrefabs, GetLootSpecs,
  GetLootSurfaceOffset, SanitizeAssetName, and the LootSpec struct.
  Adding a new loot type now means editing GetLootSpecs in this file
  and nothing else.

Main FriendSlopSceneBuilder.cs goes from 1297 to 1092 lines. Still
above the 400-line target - the remaining content is in-scene Create*
helpers (CreateLighting, CreateLevel, CreateLaunchpad, CreateShipInterior,
CreateRuntimeBootstrapper, CreateRuntimeUi) and Repair-path Ensure*
methods. Those are coherent enough as a unit to defer to a follow-up
commit; the prefab builders were the most independent extractions and
the highest-priority for "adding a new prefab type" workflows.

Verified with dotnet build on Editor and Runtime: 0 errors.
