+++ Space Game/Assets/Scripts/Editor/Builders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 45321fff50fb4419b9e200ea210d055c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
