+++ Space Game/Assets/Scripts/Editor/Builders/FriendSlopSceneBuilder.LootPrefab.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d372194a8a824fc7b4a10ce4a55de5e0
