+++ Space Game/Assets/Scripts/Editor/Builders/FriendSlopSceneBuilder.Materials.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9436d132f67e46539c9a6907b4676899
