+++ Space Game/Assets/Scripts/Editor/Builders/FriendSlopSceneBuilder.MonsterPrefab.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 941f64c7974043afb11f24f2e6ddd0be
