+++ Space Game/Assets/Scripts/Editor/Builders/FriendSlopSceneBuilder.PlayerPrefab.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 98c9d70fa0e343d38cdb5af1ea97be61
